+++ models/MySqlModels/postModel.js
@@ -5,7 +5,7 @@ export const PostTable = () => {
             user_id INT NOT NULL,
         title VARCHAR(100),
         content_path VARCHAR(100) DEFAULT "default",
-        created_time INT NOT NULL,
+        created_time bigint NOT NULL,
         FOREIGN KEY (user_id) REFERENCES user(Id) ON DELETE CASCADE
         )
         `, (err, result) => {
